Add files via upload
Fixed the design bugs when automation of locators are failing because of design changes.
E-commerce Website (Which I chosen) is a dynamic website, it changes it's design and contents, so as locators are also changing because of that Failing to meet certain requirements. If any changes happens to the website, will change the code as per requirements.